type migration for mandelbrot canvas

+++ app/src/main/java/jetsetilly/mandelbrot/Mandelbrot/MandelbrotCanvas.java
@@ -2,7 +2,7 @@ package jetsetilly.mandelbrot.Mandelbrot;
 
 public interface MandelbrotCanvas {
     void startDraw(Mandelbrot.RenderMode render_mode);
-    void drawPoint(float dx, float dy, int iteration);
+    void drawPoint(int dx, int dy, int iteration);
     void update();
     void endDraw();
     void cancelDraw();
