+++ app/src/main/java/jetsetilly/mandelbrot/RenderCanvas/BufferPixels.java
@@ -1,7 +1,6 @@
 package jetsetilly.mandelbrot.RenderCanvas;
 
 import android.graphics.Bitmap;
-import android.os.AsyncTask;
 
 import java.util.Timer;
 import java.util.TimerTask;
@@ -71,7 +70,7 @@ public class BufferPixels implements Buffer {
     }
 
     @Override
-    public void pushDraw(float cx, float cy, int iteration) {
+    public void pushDraw(int cx, int cy, int iteration) {
         // figure out which colour to use
         int palette_entry = iteration;
